terraform_data_gcp/projects/lucas-datalake/cloud-function/function-person-events/requirements.txt
pandas
Faker
google-cloud-pubsub
simplejson
google-cloud-storage